fixed total to int

+++ lib/application/app.dart
@@ -189,7 +189,7 @@ class _AddInputState extends State<AddInput> {
   String? carbRatioSelected;
   ValueNotifier<double> glucoseLevelNotifier = ValueNotifier<double>(0.0);
   ValueNotifier<double> carbsTotalNotifier = ValueNotifier<double>(0.0);
-  ValueNotifier<double> bolusCalculation = ValueNotifier<double>(0);
+  ValueNotifier<int> bolusCalculation = ValueNotifier<int>(0);
   void updateBolusCalculation() {
     if (_GlucoseController.text.isNotEmpty) {
       /* &&
@@ -250,7 +250,7 @@ class _AddInputState extends State<AddInput> {
                       children: [
                         InkWell(
                           onTap: () {
-                            ValueListenableBuilder<double>(
+                            ValueListenableBuilder<int>(
                               valueListenable: bolusCalculation,
                               builder: (context, value, child) {
                                 return Text(
@@ -340,7 +340,7 @@ class _AddInputState extends State<AddInput> {
                                 children: [
                                   Column(
                                     children: [
-                                      ValueListenableBuilder<double>(
+                                      ValueListenableBuilder<int>(
                                         valueListenable: bolusCalculation,
                                         builder: (context, value, child) {
                                           return Text(
